+++ tests/evaluation/test_event_based_metrics.py
@@ -1,7 +1,7 @@
 
 import pytest
 import numpy as np
-from dtaianomaly.evaluation.event_wise_metrics import EventWisePrecision, EventWiseRecall, EventWiseFBeta, _make_intervals, _compute_event_wise_metrics
+from dtaianomaly.evaluation.event_wise_metrics import EventWisePrecision, EventWiseRecall, EventWiseFBeta, _compute_event_wise_metrics
 
 
 class TestInitialization:
@@ -22,28 +22,6 @@ class TestInitialization:
     def test_valid_beta(self, beta):
         metric = EventWiseFBeta(beta=beta)
         assert metric.beta == beta
-
-
-class TestMakeIntervals:
-
-    @pytest.mark.parametrize('value,expected', [
-        ([], []),
-        ([0, 0, 0], []),
-        ([1, 1, 1], [(0, 2)]),
-        ([0, 1, 1, 0], [(1, 2)]),
-        ([1, 1, 0, 1], [(0, 1), (3, 3)]),
-        ([0, 1, 0, 1, 1, 0], [(1, 1), (3, 4)]),
-        ([1, 0, 0, 1, 1, 1], [(0, 0), (3, 5)]),
-        ([1], [(0, 0)]),
-        ([0], []),
-        ([0, 1, 1, 1, 0, 0, 1, 0], [(1, 3), (6, 6)]),
-    ])
-    def test_make_intervals(self, value, expected):
-        start, end = _make_intervals(np.array(value))
-        assert start.shape[0] == end.shape[0] == len(expected)
-        for s, e, (s_expected, e_expected) in zip(start, end, expected):
-            assert s == s_expected
-            assert e == e_expected
 
 
 class TestComputeEventWiseMetrics: